backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0
redis==5.0.4
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
langgraph==0.1.4
tiktoken==0.7.0
chromadb==0.5.0
langchain-chroma==0.1.1
structlog==24.2.0
sentry-sdk[fastapi]==2.3.1
prometheus-fastapi-instrumentator==7.0.0
python-dotenv==1.0.1
boto3==1.34.101
pillow==10.3.0
pypdf==4.2.0
python-docx==1.1.2
tenacity==8.3.0
reportlab==4.2.0
sse-starlette==2.1.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
factory-boy==3.3.0
